Attempting to add Vue to make this more dynamic. Probably will soon include Tabulator as well.
Flask can still be the base framework but if the data updating is going to be dynamic (YOY <--> QOQ, changing starting year and quarter), why use plain AJAX and figure out how to map the new data to HTML tags more manually when I could use the power of Vue and Tabulator? The work will be about the same amount, but instead of figuring out manual stuff, I'll be learning frameworks. Seems like the right path forward.